telemetry-client/src/windows/requirements.txt
grpcio
asyncio
time
pydantic
socket
subprocess
platform
psutil
json
typing
functools
PyYAML
grpcio-tools
pywin32
ssl